# docs/openapi.yaml
openapi: 3.1.0
info:
  title: leCRM API
  version: 0.1.0
  summary: Workspace-scoped CRM REST API + service-token administration
  description: |
    leCRM v0 HTTP API — Sprint 7 surface.

    Workspace identity is resolved from the request subdomain (e.g.
    `https://acme.lecrm.fr/v1/contacts`). All endpoints under `/v1` are
    workspace-scoped; the workspace tenant schema is derived from the
    `workspaces.role_name` column (per ADR-009 §2.1).

    Two authentication modes are supported:
      * Session cookie (browser) — set by the OIDC callback under
        `/auth/callback`. Cookie name `lecrm_session_<workspace>`.
      * Bearer token (programmatic / MCP / chatboting connector) —
        sent as `Authorization: Bearer lecrm_<workspace>_<random>`.
        Tokens are minted via `POST /v1/workspace/tokens` and stored
        as argon2id hashes (ADR-009 §4.1).

    Idempotency: every `POST` accepts an optional `Idempotency-Key`
    header. Replays within 24h return the cached response with the
    `Idempotency-Replayed: true` response header.

    Pagination: list endpoints return a `next_cursor` (opaque base64
    URL-safe value) and a `has_more` boolean. Pass the `next_cursor`
    back to the same endpoint via the `cursor` query parameter to fetch
    the next page.
servers:
  - url: https://{workspace}.lecrm.fr
    description: Production
    variables:
      workspace:
        default: acme
        description: Workspace subdomain slug
...
    variables:
...
paths:
...
  /v1/workspace/tokens:
    get:
      tags: [workspace, tokens]
      summary: List workspace service tokens (hashes never returned)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ServiceToken' }
    post:
      tags: [workspace, tokens]
      summary: |
        Create a new service token. The plaintext token is returned
        ONCE in the response and is never persisted.
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceTokenCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceTokenCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/workspace/tokens/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    delete:
      tags: [workspace, tokens]
      summary: Revoke (delete) a service token
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lecrm_session
      description: |
        Workspace-scoped HMAC-signed session cookie set by the OIDC
        callback. Cookie name is suffixed with the workspace slug.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: lecrm_<workspace>_<random>
      description: |
        Workspace service token minted via POST /v1/workspace/tokens.
        Hashed with argon2id; only the plaintext returned at creation
        time can be presented as a Bearer credential.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 255 }
      description: |
        Opaque client-supplied token; replays within 24h return the
        cached response with `Idempotency-Replayed: true`.
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ServiceToken:
      type: object
      required: [id, name, actor_type, scopes, created_at]
      properties:
        id:           { type: string, format: uuid }
        name:         { type: string }
        actor_type:   { type: string, enum: [human_api, mcp_agent, internal_service, connector] }
        scopes:       { type: array, items: { type: string } }
        expires_at:   { type: [string, 'null'], format: date-time }
        last_used_at: { type: [string, 'null'], format: date-time }
        created_at:   { type: string, format: date-time }
    ServiceTokenCreate:
      type: object
      required: [name, actor_type]
      properties:
        name:       { type: string, minLength: 1, maxLength: 200 }
        actor_type: { type: string, enum: [human_api, mcp_agent, internal_service, connector] }
        scopes:     { type: array, items: { type: string }, default: ["*"] }
        expires_at: { type: string, format: date-time }
    ServiceTokenCreated:
      type: object
      required: [token, service_token]
      description: |
        The plaintext `token` is returned ONCE — store it now; the
        server only persists the argon2id hash.
      properties:
        token:         { type: string, description: "Plaintext, format: lecrm_<workspace>_<random>" }
        service_token: { $ref: '#/components/schemas/ServiceToken' }
security:
  - sessionCookie: []
  - bearerToken: []